+++ packages/engine-p5/index.js
@@ -30,7 +30,7 @@ export const run = (functionName, func, values, config) => {
     sketch.frameRate(mechanic.settings.frameRate);
     return func.handler({
       inputs: mechanic.values,
-      ...mechanic.callbacksForEngine({
+      ...mechanic.callbacksForDesignFunction({
         drawLoop: () => {
           throw new Error(
             "drawLoop is not supported in engine p5, as p5 provides its own draw loop. Use sketch.setup and sketch.draw instead."