+++ packages/engine-svg/index.js
@@ -28,7 +28,7 @@ export const run = (functionName, func, values, config) => {
 
   func.handler({
     inputs: mechanic.values,
-    ...mechanic.callbacksForEngine()
+    ...mechanic.callbacksForDesignFunction()
   });
 
   return mechanic;